Time: 3 ms (54.79%), Space: 9.2 MB (16.85%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.c
@@ -1,7 +1,7 @@
 
 
 int reverse(int x){ 
-    long result=0;
+    long long result=0;
     int temp=x;
     while(temp!=0){
         result=(result*10)+(temp%10);
